#11 update test and importing method
