+++ lib/video.js
@@ -1,4 +1,4 @@
-import duration from './video/duration'
+import duration from './video/duration.js'
 
 const video = { duration }
