+++ test/video.js
@@ -1,0 +1,4 @@
+import video from '../lib/video.js'
+import videoDurationTest from './video/shared.js'
+
+videoDurationTest(video.duration)
